requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
spacy>=3.7.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
google-cloud-vision>=3.4.0
pandas>=2.0.0

